Refactor PostnonlinearCamSimulation to AdditiveModel
Renamed and refactored PostnonlinearCamSimulation into AdditiveModel for clarity and reusability. Updated parameter names, comments, and logic for consistency with the new additive model semantics. Adjusted documentation, constants, and method signatures across relevant components.

+++ tetrad-lib/src/main/java/edu/cmu/tetrad/util/StatUtils.java
@@ -2195,7 +2195,7 @@ public final class StatUtils {
      * @param c a int
      * @return a double
      */
-    public static double factorial(int c) {
+    public static long factorial(int c) {
         if (c < 0) throw new IllegalArgumentException("Can't take the factorial of a negative number: " + c);
         if (c == 0) return 1;
         return c * StatUtils.factorial(c - 1);
